chore: Refactor post contribution endpoint and services
Introduce a `Contribution` __DTO__ and an `Assembler` service.
> https://martinfowler.com/eaaCatalog/dataTransferObject.html

There are not many sanity checks for the moment.
And contrary to what was discuss before, this implementation is finally tightly coupled to the extension instead of being generic.

The also introduce 2 new "visibility" flags:
- draft
- question
Might not be the best chocie to put these new status under "visibility", but this was the simplest solution...

feat(contribution): Create `DomainName` automatically

fix contributor choice and message

fix: Handle case where there's no questioned contributor whereas it's still a question