fix: #304 using random filename generation
This fix prevents people|bots from leeching avatars
with a simple loop on integers.
It could be improved by using a third-party unique identifier generator
such as Ramsey's UUID package.

Another approach would be to orchestrate VichUploader and EasyAdmin
so that somehow Vich waits for the entity to be persisted once
before handling files, and so we could keep numerical ids in filenames.