feat(domain): add aliases to a domain name
Aliases are mapped to a simple php array, there is a little hack to overcome an issue with Doctrine's `simple_array` which maps the `[]` to `null` in database, hence the oddities you can see in the `DomainName` entity.

This also update the "big" regexp we expose with the possible aliases of domain. (see updated tests for example of generated regexp).

The same validation rule as a `Domain` `name` has been added to each `aliases` element but both seems to be ignored by __ZiziAdmin__...

> This fix issue #429 :kiss: