feat(matching-context): expose list of domains in API
Exposes the list of domains on which a matching context is applicable. It also exposes the *path* part of the regexp uses in mathching context...

Some current and previous vocabulary seems dodgy... maybe a `DomainName` should be renamed to `Domain` so we can use `domainName` only for the actual string version of the domain name...

Also `urlPathRegex` is not truly a *path* __regex__ since the user can choose to match all domains... is that still supposed to be possible?

> fix #446